Extract template code from gauge buffers, add adjoint fields as implementation.
Since the code for the buffers is very much duplicate for different types,
I decided to create preprocessor templates from it. This way, the source code
for the specific implementations is very shallow. If changes need to be made,
they can be made to the templates directly and are propagated throughout all
the buffer structures. This should be helpful in avoiding repetitive fixes if
bugs are discovered and guarantee a uniform interface.